+++ packages/analytics/dashboard-renderer/src/constants.ts
@@ -1,3 +1,5 @@
 // Default tile height and width - approximated from the figma design.
 export const DEFAULT_TILE_HEIGHT = 170
 export const INJECT_QUERY_PROVIDER = 'analytics-query-provider'
+export const ENTITY_ID_TOKEN = '{entity-id}'
+export const CP_ID_TOKEN = '{cp-id}'