+++ packages/entities/entities-plugins/src/components/PluginSelect.vue
@@ -69,8 +69,12 @@
             <p class="tab-description">
               {{ t('plugins.select.tabs.kong.description') }}
             </p>
+
             <PluginSelectGrid
               :config="config"
+              :hide-highlighted-plugins="filter.length > 0"
+              :highlighted-plugins="highlightedPlugins"
+              :highlighted-plugins-title="props.highlightedPluginsTitle"
               :navigate-on-click="navigateOnClick"
               :plugin-list="filteredPlugins"
               :plugins-per-row="pluginsPerRow"
@@ -105,6 +109,9 @@
       <PluginSelectGrid
         v-else
         :config="config"
+        :hide-highlighted-plugins="filter.length > 0"
+        :highlighted-plugins="highlightedPlugins"
+        :highlighted-plugins-title="props.highlightedPluginsTitle"
         :navigate-on-click="navigateOnClick"
         :plugin-list="filteredPlugins"
         :plugins-per-row="pluginsPerRow"
@@ -208,6 +215,20 @@ const props = defineProps({
     type: Number,
     default: 4,
   },
+  /**
+   * Ids of plugins to show in the highlighted plugins group
+   */
+  highlightedPluginIds: {
+    type: Array as PropType<string[]>,
+    default: () => [],
+  },
+  /**
+   * Title for the highlighted plugins group
+   */
+  highlightedPluginsTitle: {
+    type: String,
+    default: '',
+  },
 })
 
 const emit = defineEmits<{
@@ -235,6 +256,21 @@ const { axiosInstance } = useAxios({
   headers: props.config.requestHeaders,
 })
 
+const flattenPluginMap = computed(() => {
+  if (!pluginsList.value) {
+    return {}
+  }
+
+  return Object.entries(pluginsList.value)
+    .filter(([group]) => group !== PluginGroup.CUSTOM_PLUGINS)
+    .reduce((m, [, plugins]) => {
+      for (const plugin of plugins) {
+        m[plugin.id] = plugin
+      }
+      return m
+    }, {} as Record<string, PluginType>)
+})
+
 const filteredPlugins = computed((): PluginCardList => {
   if (!pluginsList.value) {
     return {}
@@ -254,6 +290,18 @@ const filteredPlugins = computed((): PluginCardList => {
   }
 
   return results
+})
+
+const highlightedPlugins = computed(() => {
+  return props.highlightedPluginIds.reduce((plugins, id) => {
+    const plugin = flattenPluginMap.value[id]
+
+    if (plugin) {
+      plugins.push(plugin)
+    }
+
+    return plugins
+  }, [] as PluginType[])
 })
 
 const hasFilteredResults = computed((): boolean => {
@@ -392,7 +440,7 @@ const fetchEntityPluginsUrl = computed((): string => {
 })
 
 const onTabsChange = (hash: string) => {
-  router.replace({ hash })
+  router.replace({ hash, query: route.query })
 }
 
 // race condition between fetch of available plugins and setting