+++ packages/analytics/analytics-config-store/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.2.0...@kong-ui-public/analytics-config-store@0.3.0) (2024-03-19)
+
+
+### Features
+
+* **analytics:** move to v2 config endpoint [MA-2730] ([#1257](https://github.com/Kong/public-ui-components/issues/1257)) ([e61e42f](https://github.com/Kong/public-ui-components/commit/e61e42fddb221a8e04f7faa18aa965bf7584d0e3))
+
+
+### BREAKING CHANGES
+
+* **analytics:** new config endpoint; types updated
+
+The following components are updated:
+
+- AnalyticsConfigCheck, and the associated store
+- MetricsProviderInternal
+- AnalyticsBridge (updates type that implementations must adhere to)
+- DashboardRenderer (sandbox and tests)
+
+
+
+
+
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.1.2...@kong-ui-public/analytics-config-store@0.2.0) (2024-03-14)
+
+
+### Features
+
+* **analytics:** add `AnalyticsConfigCheck` component [MA-2715] ([#1237](https://github.com/Kong/public-ui-components/issues/1237)) ([e6ec171](https://github.com/Kong/public-ui-components/commit/e6ec1717e34a1d8ba994bf9dc4948cdef21776f2))
+
+
+
+
+
+## [0.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.1.1...@kong-ui-public/analytics-config-store@0.1.2) (2024-03-13)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-config-store
+
+
+
+
+
 ## [0.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-config-store@0.1.0...@kong-ui-public/analytics-config-store@0.1.1) (2024-03-12)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-config-store