+++ packages/entities/entities-consumer-credentials/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.77](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@2.1.76...@kong-ui-public/entities-consumer-credentials@2.1.77) (2024-03-19)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
+## [2.1.76](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@2.1.75...@kong-ui-public/entities-consumer-credentials@2.1.76) (2024-03-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
+## [2.1.75](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@2.1.74...@kong-ui-public/entities-consumer-credentials@2.1.75) (2024-03-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
+## [2.1.74](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@2.1.73...@kong-ui-public/entities-consumer-credentials@2.1.74) (2024-03-13)
+
+**Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials
+
+
+
+
+
 ## [2.1.73](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-consumer-credentials@2.1.72...@kong-ui-public/entities-consumer-credentials@2.1.73) (2024-03-13)
 
 **Note:** Version bump only for package @kong-ui-public/entities-consumer-credentials