+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.5.1...@kong-ui-public/entities-upstreams-targets@2.5.2) (2024-03-19)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [2.5.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.5.0...@kong-ui-public/entities-upstreams-targets@2.5.1) (2024-03-18)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+# [2.5.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.4.5...@kong-ui-public/entities-upstreams-targets@2.5.0) (2024-03-18)
+
+
+### Features
+
+* **jsonYamlForms:** remove json yaml milestone 2 ff for all entity forms post GA [KHCP-9971] ([#1259](https://github.com/Kong/public-ui-components/issues/1259)) ([109cd54](https://github.com/Kong/public-ui-components/commit/109cd544c1d706641c9cbf3df3532e85765f3a3a))
+
+
+
+
+
+## [2.4.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.4.4...@kong-ui-public/entities-upstreams-targets@2.4.5) (2024-03-13)
+
+
+### Bug Fixes
+
+* **upstream-list:** use exact match search on upstream list [KHCP-11054] ([#1230](https://github.com/Kong/public-ui-components/issues/1230)) ([d7266c3](https://github.com/Kong/public-ui-components/commit/d7266c38c4260eb78bf2b721e3d9f21a92cc4529))
+
+
+
+
+
 ## [2.4.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@2.4.3...@kong-ui-public/entities-upstreams-targets@2.4.4) (2024-03-13)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets